Make it so the filesystem supports files bigger than 4gb in 32bit builds also, not just x64. Reads/Writes are still restricted to 32bit in both x86 and x64.

+++ Pulp/Core/DiskFileAsync.cpp
@@ -17,12 +17,12 @@ XDiskFileAsync::~XDiskFileAsync()
 }
 
 
-XFileAsyncOperation XDiskFileAsync::readAsync(void* pBuffer, size_t length, size_t position)
+XFileAsyncOperation XDiskFileAsync::readAsync(void* pBuffer, size_t length, uint64_t position)
 {
 	return XFileAsyncOperation(file_.readAsync(pBuffer, length, position), pBuffer);
 }
 
-XFileAsyncOperation XDiskFileAsync::writeAsync(const void* pBuffer, size_t length, size_t position)
+XFileAsyncOperation XDiskFileAsync::writeAsync(const void* pBuffer, size_t length, uint64_t position)
 {
 	return XFileAsyncOperation(file_.writeAsync(pBuffer, length, position), pBuffer);
 }
@@ -37,12 +37,12 @@ size_t XDiskFileAsync::WaitUntilFinished(const XFileAsyncOperation& operation)
 	return numBytes;
 }
 
-size_t XDiskFileAsync::remainingBytes(void) const
+uint64_t XDiskFileAsync::remainingBytes(void) const
 {
 	return file_.remainingBytes();
 }
 
-void XDiskFileAsync::setSize(size_t numBytes)
+void XDiskFileAsync::setSize(int64_t numBytes)
 {
 	return file_.setSize(numBytes);
 }
