+++ Pulp/Core/win32/OsFileAsync.h
@@ -14,19 +14,19 @@ struct OsFileAsync
 	OsFileAsync(const wchar_t* path, IFileSys::fileModeFlags mode);
 	~OsFileAsync(void);
 
-	XOsFileAsyncOperation readAsync(void* pBuffer, size_t length, size_t position);
-	XOsFileAsyncOperation writeAsync(const void* pBuffer, size_t length, size_t position);
+	XOsFileAsyncOperation readAsync(void* pBuffer, size_t length, uint64_t position);
+	XOsFileAsyncOperation writeAsync(const void* pBuffer, size_t length, uint64_t position);
 
-	size_t tell(void) const;
-	size_t remainingBytes(void) const;
-	void setSize(size_t numBytes);
+	uint64_t tell(void) const;
+	uint64_t remainingBytes(void) const;
+	void setSize(int64_t numBytes);
 
 	bool valid(void) const;
 
 	static XFileStats& fileStats(void);
 
 private:
-	void seek(size_t position, IFileSys::SeekMode::Enum origin);
+	void seek(int64_t position, IFileSys::SeekMode::Enum origin);
 
 private:
 	IFileSys::fileModeFlags mode_;
